requirements.txt
azure-storage-blob
pyyaml
